DAH-Capstone/POC1/requirements.txt
torch
transformers
faiss-cpu
nltk
pymupdf
openai>=1.0.0
numpy